Feat: split section of About tab into resources tab
