+++ src/config/routes.ts
@@ -1,7 +1,8 @@
 import IRoute from '../interfaces/IRoute';
+import HomePage from '../pages/home';
 import BuildingsPage from '../pages/buildings';
 import AboutPage from '../pages/about';
-import HomePage from '../pages/home';
+import ResourcesPage from '../pages/resources';
 import SavedHomesPage from '../pages/saved-homes';
 import DashboardPage from '../auth_components/Dashboard'
 
@@ -22,6 +23,12 @@ const routes: IRoute[] = [
     path: '/about',
     name: 'About Page',
     component: AboutPage,
+    exact: true
+  },
+  {
+    path: '/resources',
+    name: 'Resources Page',
+    component: ResourcesPage,
     exact: true
   },
   {
